Added possibility to check for formats.
I added a step that checks whether a property has a certain format.
The format can be a simple name, or a name with an additional argument.
This gives the option to use formatters like 'even' or 'regex ^[a-z]+$'.

Currently, one placeholder format ('even') is defined. This needs to be
replaced with a base set.

I also added an action that can be used to load custom formatters.